[libc++] regarding frederick-vs-ja's feedback, flat_* dedup helper and libc++-specific tests
