+++ libcxx/include/__flat_map/flat_map.h
@@ -938,17 +938,18 @@ private:
     return ranges::adjacent_find(__key_container, __greater_or_equal_to) == ranges::end(__key_container);
   }
 
-  template <class _Iter, class _Sent, class _Predicate>
-  _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 static _Iter
-  __deduplicate_sorted(_Iter __first, _Sent __last, _Predicate __pred) {
-    __first = ranges::adjacent_find(__first, __last, __pred);
+  template <class _Iter, class _Sent>
+  _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 _Iter __deduplicate_sorted(_Iter __first, _Sent __last) {
+    __first = ranges::adjacent_find(__first, __last, [this](const auto& __x, const auto& __y) -> bool {
+      return !__compare_(std::get<0>(__x), std::get<0>(__y));
+    });
     if (__first == __last) {
       return __first;
     }
 
     _Iter __i = __first;
     for (++__i; ++__i != __last;) {
-      if (!__pred(*__first, *__i)) {
+      if (!__compare_(std::get<0>(*__first), std::get<0>(*__i))) {
         *++__first = _IterOps<_RangeAlgPolicy>::__iter_move(__i);
       }
     }
@@ -961,7 +962,7 @@ private:
   _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 void __sort_and_unique() {
     auto __zv = ranges::views::zip(__containers_.keys, __containers_.values);
     ranges::sort(__zv, __compare_, [](const auto& __p) -> decltype(auto) { return std::get<0>(__p); });
-    auto __dup_start = __deduplicate_sorted(__zv.begin(), __zv.end(), __key_not_less(__compare_));
+    auto __dup_start = __deduplicate_sorted(__zv.begin(), __zv.end());
     auto __dist      = ranges::distance(__zv.begin(), __dup_start);
     __containers_.keys.erase(__containers_.keys.begin() + __dist, __containers_.keys.end());
     __containers_.values.erase(__containers_.values.begin() + __dist, __containers_.values.end());
@@ -996,7 +997,7 @@ private:
       }
       ranges::inplace_merge(__zv.begin(), __zv.begin() + __append_start_offset, __end, __compare_key);
 
-      auto __dup_start = __deduplicate_sorted(__zv.begin(), __zv.end(), __key_not_less(__compare_));
+      auto __dup_start = __deduplicate_sorted(__zv.begin(), __zv.end());
       auto __dist      = ranges::distance(__zv.begin(), __dup_start);
       __containers_.keys.erase(__containers_.keys.begin() + __dist, __containers_.keys.end());
       __containers_.values.erase(__containers_.values.begin() + __dist, __containers_.values.end());
@@ -1152,15 +1153,6 @@ private:
   containers __containers_;
   _LIBCPP_NO_UNIQUE_ADDRESS key_compare __compare_;
 
-  struct __key_not_less {
-    _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 __key_not_less(key_compare __c) : __comp_(__c) {}
-    _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 bool
-    operator()(const_reference __x, const_reference __y) const {
-      // __x is always before __y in sorted ranges.
-      return !__comp_(std::get<0>(__x), std::get<0>(__y));
-    }
-    key_compare __comp_;
-  };
 };
 
 template <class _KeyContainer, class _MappedContainer, class _Compare = less<typename _KeyContainer::value_type>>
