+++ libcxx/include/__flat_set/flat_set.h
@@ -683,17 +683,18 @@ private:
     return ranges::adjacent_find(__key_container, __greater_or_equal_to) == ranges::end(__key_container);
   }
 
-  template <class _Iter, class _Sent, class _Predicate>
-  _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 static _Iter
-  __deduplicate_sorted(_Iter __first, _Sent __last, _Predicate __pred) {
-    __first = ranges::adjacent_find(__first, __last, __pred);
+  template <class _Iter, class _Sent>
+  _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 _Iter __deduplicate_sorted(_Iter __first, _Sent __last) {
+    __first = ranges::adjacent_find(__first, __last, [this](const auto& __x, const auto& __y) -> bool {
+      return !__compare_(__x, __y);
+    });
     if (__first == __last) {
       return __first;
     }
 
     _Iter __i = __first;
     for (++__i; ++__i != __last;) {
-      if (!__pred(*__first, *__i)) {
+      if (!__compare_(*__first, *__i)) {
         *++__first = _IterOps<_RangeAlgPolicy>::__iter_move(__i);
       }
     }
@@ -705,7 +706,7 @@ private:
   // is no invariant state to preserve
   _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 void __sort_and_unique() {
     ranges::sort(__keys_, __compare_);
-    auto __dup_start = __deduplicate_sorted(__keys_.begin(), __keys_.end(), __key_not_less(__compare_));
+    auto __dup_start = __deduplicate_sorted(__keys_.begin(), __keys_.end());
     __keys_.erase(__dup_start, __keys_.end());
   }
 
@@ -723,7 +724,7 @@ private:
       }
       ranges::inplace_merge(__keys_.begin(), __keys_.begin() + __old_size, __keys_.end(), __compare_);
 
-      auto __dup_start = __deduplicate_sorted(__keys_.begin(), __keys_.end(), __key_not_less(__compare_));
+      auto __dup_start = __deduplicate_sorted(__keys_.begin(), __keys_.end());
       __keys_.erase(__dup_start, __keys_.end());
     }
     __on_failure.__complete();
@@ -798,15 +799,6 @@ private:
   _KeyContainer __keys_;
   _LIBCPP_NO_UNIQUE_ADDRESS key_compare __compare_;
 
-  struct __key_not_less {
-    _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 __key_not_less(key_compare __c) : __comp_(__c) {}
-    _LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX26 bool
-    operator()(const_reference __x, const_reference __y) const {
-      // __x is always before __y in sorted ranges.
-      return !__comp_(__x, __y);
-    }
-    key_compare __comp_;
-  };
 };
 
 template <class _KeyContainer, class _Compare = less<typename _KeyContainer::value_type>>
